TNO-1751 PR updates
