+++ app/editor/src/features/content/list-view/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './applySortBy';
+export * from './getActions';
 export * from './getStatusText';
 export * from './makeFilter';
 export * from './queryToFilter';
