+++ libs/npm/core/src/hooks/api-editor/interfaces/IFilterSettingsModel.ts
@@ -23,5 +23,6 @@ export interface IFilterSettingsModel extends Record<string, any> {
   contentTypes: ContentTypeName[];
   tags: string[];
   sentiment: number[];
-  hasTopic: boolean;
+  hasTopic?: boolean;
+  status?: string;
 }
