+++ libs/npm/core/src/utils/elastic/generateQuery.ts
@@ -34,9 +34,10 @@ export const generateQuery = (settings: IFilterSettingsModel, query: any = {}) =
           generateRangeForArrayField('tonePools.value', settings.sentiment),
           ...generateQueryForActions(settings.actions ?? []),
           generateTextQuery(settings),
-          generateTerm('edition', settings.edition),
-          generateTerm('section', settings.section),
-          generateTerm('page', settings.page),
+          settings.edition ? generateTerm('edition', `*${settings.edition}*`) : undefined,
+          settings.section ? generateTerm('section', `*${settings.section}*`) : undefined,
+          settings.page ? generateTerm('page', `*${settings.page}*`) : undefined,
+          settings.status ? generateTerm('status', settings.status) : undefined,
           settings.hasTopic ? generateQueryForExistCheck('topics') : undefined,
         ].filter((v) => v !== undefined),
       },
